+++ sympy/matrices/tests/test_commonmatrix.py
@@ -771,87 +771,6 @@ def test_div():
     assert n/2 == ArithmeticOnlyMatrix(1, 2, [S.Half, S(2)/2])
 
 
-def test_adjugate():
-    x = Symbol('x')
-    e = DeterminantOnlyMatrix(4, 4,
-        [x, 1, 2, 3, 4, 5, 6, 7, 2, 9, 10, 11, 12, 13, 14, 14])
-
-    adj = Matrix([
-        [   4,         -8,         4,         0],
-        [  76, -14*x - 68,  14*x - 8, -4*x + 24],
-        [-122, 17*x + 142, -21*x + 4,  8*x - 48],
-        [  48,  -4*x - 72,       8*x, -4*x + 24]])
-    assert e.adjugate() == adj
-    assert e.adjugate(method='bareiss') == adj
-    assert e.adjugate(method='berkowitz') == adj
-
-    a = DeterminantOnlyMatrix(2, 3, [1, 2, 3, 4, 5, 6])
-    raises(NonSquareMatrixError, lambda: a.adjugate())
-
-
-def test_cofactor_and_minors():
-    x = Symbol('x')
-    e = DeterminantOnlyMatrix(4, 4,
-        [x, 1, 2, 3, 4, 5, 6, 7, 2, 9, 10, 11, 12, 13, 14, 14])
-
-    m = Matrix([
-        [ x,  1,  3],
-        [ 2,  9, 11],
-        [12, 13, 14]])
-    cm = Matrix([
-        [ 4,         76,       -122,        48],
-        [-8, -14*x - 68, 17*x + 142, -4*x - 72],
-        [ 4,   14*x - 8,  -21*x + 4,       8*x],
-        [ 0,  -4*x + 24,   8*x - 48, -4*x + 24]])
-    sub = Matrix([
-            [x, 1,  2],
-            [4, 5,  6],
-            [2, 9, 10]])
-
-    assert e.minor_submatrix(1, 2) == m
-    assert e.minor_submatrix(-1, -1) == sub
-    assert e.minor(1, 2) == -17*x - 142
-    assert e.cofactor(1, 2) == 17*x + 142
-    assert e.cofactor_matrix() == cm
-    assert e.cofactor_matrix(method="bareiss") == cm
-    assert e.cofactor_matrix(method="berkowitz") == cm
-
-    raises(ValueError, lambda: e.cofactor(4, 5))
-    raises(ValueError, lambda: e.minor(4, 5))
-    raises(ValueError, lambda: e.minor_submatrix(4, 5))
-
-    a = DeterminantOnlyMatrix(2, 3, [1, 2, 3, 4, 5, 6])
-    assert a.minor_submatrix(0, 0) == Matrix([[5, 6]])
-
-    raises(ValueError, lambda:
-        DeterminantOnlyMatrix(0, 0, []).minor_submatrix(0, 0))
-    raises(NonSquareMatrixError, lambda: a.cofactor(0, 0))
-    raises(NonSquareMatrixError, lambda: a.minor(0, 0))
-    raises(NonSquareMatrixError, lambda: a.cofactor_matrix())
-
-
-def test_charpoly():
-    x, y = Symbol('x'), Symbol('y')
-    z, t = Symbol('z'), Symbol('t')
-
-    from sympy.abc import a,b,c
-
-    m = DeterminantOnlyMatrix(3, 3, [1, 2, 3, 4, 5, 6, 7, 8, 9])
-
-    assert eye_Determinant(3).charpoly(x) == Poly((x - 1)**3, x)
-    assert eye_Determinant(3).charpoly(y) == Poly((y - 1)**3, y)
-    assert m.charpoly() == Poly(x**3 - 15*x**2 - 18*x, x)
-    raises(NonSquareMatrixError, lambda: Matrix([[1], [2]]).charpoly())
-
-    n = DeterminantOnlyMatrix(4, 4, [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0])
-    assert n.charpoly() == Poly(x**4, x)
-
-    n = DeterminantOnlyMatrix(4, 4, [45, 0, 0, 0, 0, 23, 0, 0, 0, 0, 87, 0, 0, 0, 0, 12])
-    assert n.charpoly() == Poly(x**4 - 167*x**3 + 8811*x**2 - 173457*x + 1080540, x)
-
-    n = DeterminantOnlyMatrix(3, 3, [x, 0, 0, a, y, 0, b, c, z])
-    assert n.charpoly() == Poly(t**3 - (x+y+z)*t**2 + t*(x*y+y*z+x*z) - x*y*z , t)
-
 # ReductionsOnlyMatrix tests
 def test_row_op():
     e = eye_Reductions(3)
